Add useAffirmation and useColorPicker custom hooks w passing test for latter
